requirements.txt
google-genai
streamlit
pandas
pydub
PyAudio
setuptools
python-dotenv
streamlit-webrtc
yapper-tts